% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_datasets.R
\name{available_datasets}
\alias{available_datasets}
\title{Available datasets via CARDdealr}
\usage{
available_datasets()
}
\value{
A \code{data.frame} of datasets available. The \code{accessor} column
contains the name of the function that returns that dataset.
}
\description{
Available datasets via CARDdealr
}
\examples{
available_datasets()

}
