% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{cdr_get_cache}
\alias{cdr_get_cache}
\title{get the BiocFileCache for CARDdealr}
\usage{
cdr_get_cache(cache = rappdirs::user_cache_dir(appname = "CARDdealr"))
}
\arguments{
\item{cache}{character(1), the path to the cache directory. See \code{\link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache()}}.}
}
\value{
a \code{BiocFileCache} instance
}
\description{
Set up or get location of BiocFileCache. By default,
this uses the name BiocFileCache.
}
\keyword{internal}
