% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_carddealr_logger}
\alias{get_carddealr_logger}
\title{The CARDdealr logger}
\usage{
get_carddealr_logger()
}
\value{
a \link[lgr:LoggerGlue]{lgr::LoggerGlue} instance
}
\description{
The CARDdealr logger
}
\examples{

lg <- get_carddealr_logger()

lg$info('this is a log entry')


}
