% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_acr_lung_cancer_screening_data.R
\name{src_acr_lung_cancer_screening_data}
\alias{src_acr_lung_cancer_screening_data}
\title{American College of Radiology Lung Cancer Screening Registry (LCSR) dataset}
\source{
\itemize{
\item \href{https://www.acr.org/Clinical-Resources/Lung-Cancer-Screening-Resources/LCS-Locator-Tool}{Data Source}
}
}
\usage{
src_acr_lung_cancer_screening_data(...)
}
\arguments{
\item{\dots}{passed to \code{\link[readr:read_delim]{readr::read_csv()}} and useful for limiting
the number of rows read for testing or glimpsing data.}
}
\value{
A data frame (data.frame) containing data pulled from LCSR.
}
\description{
\code{src_lung_cancer_screening_data()} returns a dataframe of actively reporting sites in the
lung cancer screening registry (LCSR).
}
\details{
The ACR Lung Cancer Screening Registry (LCSR) is designed to systematically audit
the quality of interpretation of screening lung CT exams. The registry is based on the
ACR Lung Imaging Reporting and Data System (Lung-RADS), which is the product of the
ACR Lung Cancer Screening Committee subgroup on Lung-RADS. This Lung-RADS system is
a quality assurance tool designed to standardize lung cancer screening CT reporting
and management recommendations, reduce confusion in lung cancer screening CT
interpretations and facilitate outcome monitoring. The ACR LCSR will capture Lung-RADS
recommendations and monitor and compare appropriate use of Lung-RADS#'
}
\examples{
# example code


lcs = src_acr_lung_cancer_screening(nrows=1000)

colnames(lcs)

dplyr::glimpse(lcs)

}
\seealso{
Other data accessors: 
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
