% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_bls_unemployment.R
\name{src_bls_unemployment_data}
\alias{src_bls_unemployment_data}
\title{Bureau of Labor Statistics Unemployment Data}
\usage{
src_bls_unemployment_data(...)
}
\arguments{
\item{\dots}{passed to \link[readr:read_delim]{readr::read_tsv} just in case....}
}
\description{
Bureau of Labor Statistics Unemployment Data
}
\examples{

bls_unemployment = src_bls_unemployment_data()

colnames(bls_unemployment)

dplyr::glimpse(bls_unemployment)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_brfss}()},
\code{\link{src_epa_superfund_data}()}
}
\concept{data accessors}
