% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_cdc_places_data.R
\name{src_cdc_places_data}
\alias{src_cdc_places_data}
\title{CDC PLACES}
\source{
\itemize{
\item \href{https://www.cdc.gov/places/measure-definitions/index.html}{Data Source}
}
}
\usage{
src_cdc_places_data(level = "county", year = 2023, ...)
}
\arguments{
\item{level}{Character string for desired geographic level. Available at 'county',
"tract", or 'zcta'. Defaults to 'county'.}

\item{year}{The year of the CDC PLACES estimates. Estimates are available from
2020 through 2023. Defaults to 2023.}
}
\value{
A data frame (data.frame) containing data pulled from the CDC PLACES API.
}
\description{
\code{src_cdc_places_data()} returns a dataframe of estimates from CDC PLACES.
}
\details{
PLACES is a collaboration between CDC, the Robert Wood Johnson Foundation,
and the CDC Foundation. PLACES provides health data for small areas across
the country. This allows local health departments and jurisdictions, regardless
of population size and rurality, to better understand the burden and
geographic distribution of health measures in their areas and assist them
in planning public health interventions. PLACES provides model-based,
population-level analysis and community estimates of health measures to all
counties, places (incorporated and census designated places), census tracts,
and ZIP Code Tabulation Areas (ZCTAs) across the United States.
}
\examples{
# example code

places = src_cdc_places_data(level = 'tract', nrows = 500)

colnames(places)

dplyr::glimpse(places)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_epa_tri_data}()},
\code{\link{src_fda_mammography_data}()},
\code{\link{src_hrsa_facility_data}()},
\code{\link{src_nppes_gi_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
