% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_epa_ejscreen_data.R
\name{src_epa_ejscreen_data}
\alias{src_epa_ejscreen_data}
\title{EPA Environmental Justice Screening (EJScreen) Dataset}
\source{
\itemize{
\item \href{https://www.epa.gov/ejscreen/download-ejscreen-data}{Data Source}
}
}
\usage{
src_epa_ejscreen_data(...)
}
\value{
A data frame (data.frame) containing data pulled from the EPA EJScreen dataset.
}
\description{
\code{src_epa_ejscreen_data()} returns a dataframe of census tract level environmental
data from the EPA's EJScreen dataset.
}
\details{
EJScreen is an EPA's environmental justice mapping and screening tool that
provides EPA with a nationally consistent dataset and approach for combining
environmental and demographic socioeconomic indicators. EJScreen users choose
a geographic area; the tool then provides demographic socioeconomic and environmental
information for that area. All of the EJScreen indicators are publicly-available
data. EJScreen simply provides a way to display this information and includes
a method for combining environmental and demographic indicators into EJ indexes.
}
\examples{
# example code

#function pivots data longwise, so output is 13*n_max rows
ejs = src_epa_ejscreen_data(n_max = 500)

colnames(ejs)

dplyr::glimpse(ejs)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_cdc_places_data}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_epa_tri_data}()},
\code{\link{src_fda_mammography_data}()},
\code{\link{src_hrsa_facility_data}()},
\code{\link{src_nppes_gi_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
