% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_epa_superfund_data.R
\name{src_epa_superfund_data}
\alias{src_epa_superfund_data}
\title{Environmental Protection Agency Superfund API}
\source{
\itemize{
\item \href{https://www.epa.gov/enviro/web-services}{Data Source}
}
}
\usage{
src_epa_superfund_data()
}
\value{
A data frame (data.frame) containing data pulled from the Envirofacts API.
}
\description{
\code{src_epa_superfund_data()} returns a dataframe of Superfund sites from the Envirofacts API.
}
\details{
The Envirofacts Multisystem Search integrates information from a variety of databases and
includes latitude and longitude information. Each of these databases contains information
about facilities that are required to report activity to a state or federal system.
}
\examples{
# example code

sf = src_epa_superfund_data()
dim(sf)

if(requireNamespace('ggplot2')) {
  ggplot2::ggplot(sf, ggplot2::aes(x=State)) + 
    ggplot2::geom_bar() +
    ggplot2::coord_flip()
}

sf_ky = sf |> dplyr::filter(State=='KY')
dim(sf_ky)

colnames(sf)

dplyr::glimpse(sf)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
