% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_epa_tri_data.R
\name{src_epa_tri_data}
\alias{src_epa_tri_data}
\title{EPA Toxic Release Inventory (TRI) Dataset}
\source{
\itemize{
\item \href{https://data.epa.gov/efservice/downloads/tri/mv_tri_basic_download/}{Data Source}
}
}
\usage{
src_epa_tri_data()
}
\value{
A data frame (data.frame) containing data pulled from the Envirofacts API.
}
\description{
\code{src_epa_tri_data()} returns a dataframe of facilities reporting into the Toxic
Release Inventory.
}
\details{
The Envirofacts Multisystem Search integrates information from a variety of
databases and includes latitude and longitude information. Each of these databases
contains information about facilities that are required to report activity to
a state or federal system. The TRI contains information about all facilities
releasing reportable toxic materials, identified as carcinogenic or not.
}
\examples{
# example code

tri = src_epa_tri_data()

colnames(tri)

dplyr::glimpse(tri)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_cdc_places_data}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_fda_mammography_data}()},
\code{\link{src_hrsa_facility_data}()},
\code{\link{src_nppes_gi_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
