% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_fda_mammography_data.R
\name{src_fda_mammography_data}
\alias{src_fda_mammography_data}
\title{FDA Certified Mammography Facilities Dataset}
\source{
\itemize{
\item \href{https://www.fda.gov/radiation-emitting-products/consumer-information-mqsa/search-certified-facility}{Data Source}
}
}
\usage{
src_fda_mammography_data()
}
\value{
A data frame (data.frame) containing data pulled from the Envirofacts API.
}
\description{
\code{src_fda_mammography_data()} returns a dataframe of certified mammography facilities
from the FDA Certified Mammography Facilities dataset.
}
\details{
The Mammography Quality Standards Act requires mammography facilities across the nation
to meet uniform quality standards. Congress passed this law in 1992 to assure high-quality
mammography for early breast cancer detection, which can lead to early treatment, a range of
treatment options, and increased chances of survival. Under the law, all mammography facilities
must: 1) be accredited by an FDA-approved accreditation body, 2) be certified by FDA,
or its State, as meeting the standards, 3) undergo an annual MQSA inspection,
and 4) prominently display the certificate issued by the agency. All facilities meeting this criteria
are listed in the FDA Certified Mammography Facilities dataset
}
\examples{
# example code

mam = src_fda_mammography_data()

colnames(mam)

dplyr::glimpse(mam)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_cdc_places_data}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_epa_tri_data}()},
\code{\link{src_hrsa_facility_data}()},
\code{\link{src_nppes_gi_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
