% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_hrsa_facility_data.R
\name{src_hrsa_facility_data}
\alias{src_hrsa_facility_data}
\title{Health Resources and Services Administration (HRSA) facilities}
\source{
\itemize{
\item \href{https://data.hrsa.gov/data/about?tab=DataUsage}{Data Source}
}
}
\usage{
src_hrsa_facility_data()
}
\value{
A data frame (data.frame) containing data pulled from HRSA.
}
\description{
\code{src_hrsa_facility_data()} returns a dataframe of FQHC and other HPSA facilities
from HRSA.
}
\details{
HRSA programs provide health care to people who are geographically isolated,
economically or medically vulnerable. This includes people living with HIV/AIDS,
pregnant women, mothers and their families, and those otherwise unable to access
high quality health care. HRSA also supports access to health care in rural areas,
the training of health professionals, the distribution of providers to areas where
they are needed most, and improvements in health care delivery.
}
\examples{
# example code

facs = src_hrsa_facility_data()

colnames(facs)

dplyr::glimpse(facs)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_cdc_places_data}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_epa_tri_data}()},
\code{\link{src_fda_mammography_data}()},
\code{\link{src_nppes_gi_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
