% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_nppes_gi_data.R
\name{src_nppes_gi_data}
\alias{src_nppes_gi_data}
\title{National Plan & Provider Enumeration System (NPPES)}
\source{
\itemize{
\item \href{https://npiregistry.cms.hhs.gov/api-page}{Data Source}
}
}
\usage{
src_nppes_gi_data(state = c(state.abb, "DC"))
}
\arguments{
\item{state}{Character string or vector of character strings of state postal abbreviations}
}
\value{
A data frame (data.frame) containing data pulled from the NPPES API.
}
\description{
\code{src_nppes_gi_data()} returns a dataframe of GI Providers (Gastroenterology and
Colon & Recal Surgeons) from the NPPES API.
}
\details{
Centers for Medicare & Medicaid Services CMS has developed the NPPES to assign
unique identifiers to health care providers. The National Provider Indentifier (NPI)
has been the standard identifier for all HIPAA-covered entities (health care providers)
since May 23, 2007. Small health plans were required to obtain and use
an NPI by May 23, 2008.
}
\examples{
# example code

gi = src_nppes_gi_data(state  = c('KY', 'TN'))

colnames(gi)

dplyr::glimpse(gi)

}
\seealso{
Other data accessors: 
\code{\link{src_acr_lung_cancer_screening_data}()},
\code{\link{src_bls_unemployment_data}()},
\code{\link{src_brfss}()},
\code{\link{src_epa_superfund_data}()},
\code{\link{src_epa_tri_data}()},
\code{\link{src_fda_mammography_data}()},
\code{\link{src_nppes_onco_data}()}
}
\author{
Todd Burus \href{mailto:tburus@uky.edu}{tburus@uky.edu}
}
\concept{data accessors}
